callfork: Add debug message for bug hunting
At 37C3 we experienced yate crashes that are caused
by freed forkchilds that are still in the slave
list although they should've been removed.

I suspect that this is caused by a scenario where they're
removed before they even have been added. In order to gather
evidence for this suspicion, add a log message to the place
in the code that is more or less the only occasion in
which this may happen.